Update dependencies including ESlint 9 upgrade
