--- .eslintrc.js
@@ -1,5 +1,0 @@
-"use strict";
-
-module.exports = {
-  extends: "./index.js",
-};
