+++ .husky/pre-commit
@@ -1,0 +1,2 @@
+npm test
+lint-staged
